+++ src/Collections/src/ArrayList.php
@@ -80,6 +80,14 @@ class ArrayList implements IList
     }
 
     /**
+     * @inheridoc
+     */
+    public function filter(Closure $callback): static
+    {
+        return new static(\array_filter($this->values, $callback));
+    }
+
+    /**
      * @inheritdoc
      */
     public function get(int $index): mixed
@@ -125,6 +133,14 @@ class ArrayList implements IList
     public function intersect(array $values): static
     {
         return new static(\array_values(\array_intersect($this->values, $values)));
+    }
+
+    /**
+     * @inheridoc
+     */
+    public function map(Closure $callback): static
+    {
+        return new static(\array_map($callback, $this->values));
     }
 
     /**