+++ src/Collections/src/HashSet.php
@@ -83,6 +83,14 @@ class HashSet implements ISet
     }
 
     /**
+     * @inheridoc
+     */
+    public function filter(Closure $callback): static
+    {
+        return new static(\array_filter($this->values, $callback));
+    }
+
+    /**
      * @inheritdoc
      */
     public function getIterator(): Traversable
@@ -105,6 +113,14 @@ class HashSet implements ISet
         }
 
         return new static($intersectedValues);
+    }
+
+    /**
+     * @inheridoc
+     */
+    public function map(Closure $callback): static
+    {
+        return new static(\array_map($callback, $this->values));
     }
 
     /**